Fix a bug on GeneralizedTime.
DST may add a shift on local times when no offset is given in ASN.1
object. To fix it, build time object with guessed offset, then compare it
with fresher build with Time.new, which set correct UTC offset from given
date. If objects are not equal, the latter is the one to keep.